Move SaveCode functionality to a separate class
