+++ DtoGeneratorProgram/DtoGeneratorProgram/DtoGeneratorProgram.csproj
@@ -45,6 +45,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ClassDescriptioSaver.cs" />
     <Compile Include="FinderTypes.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
